% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{ask_string_write_to_file}
\alias{ask_string_write_to_file}
\title{ask_string_write_to_file}
\usage{
ask_string_write_to_file(string)
}
\arguments{
\item{string}{}
}
\description{
ask_string_write_to_file
}
\details{
ask_string_write_to_file
}
\author{
Marc Schwenzer
}
