% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{convert_to_class}
\alias{convert_to_class}
\title{convert_to_class}
\usage{
convert_to_class(a.input, class)
}
\arguments{
\item{a.input}{a vector}

\item{class}{a class}
}
\value{
the vector transformed to class
}
\description{
convert_to_class
}
\details{
Convert a input vector to according class. (Internal helper function of reco.)
}
\author{
Marc Schwenzer
}
