% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date_convert.R
\name{date_convert}
\alias{date_convert}
\title{Converts the date column to POSIXct datetime format.}
\usage{
date_convert(dataset = data, col = colNumber, name = stringname)
}
\arguments{
\item{dataset}{A dataframe.}

\item{y}{A column number of the dataframe.}
}
\value{
The date column in proper POSIXct format.
}
\description{
Converts the date column to POSIXct datetime format.
}

