% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{get_a.df}
\alias{get_a.df}
\title{.. content for \description{} (no empty lines) ..}
\usage{
get_a.df(string)
}
\arguments{
\item{string}{}
}
\value{
a.df
}
\description{
get_a.df
}
\details{
return a.df from reco env
}
\author{
Marc Schwenzer
}
