% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_api.R
\name{get_api}
\alias{get_api}
\title{Makes a call to the Eco Service 1.7 API}
\usage{
get_api(start, method)
}
\arguments{
\item{start}{the start parameter in the call.}

\item{method}{the method being called (e.g. GetCustomers).}
}
\value{
The response of the API call is in JSON format.
}
\description{
Makes a call to the Eco Service 1.7 API
}

