% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{load_df_or_create_it}
\alias{load_df_or_create_it}
\title{load_file_or_create_it}
\usage{
load_df_or_create_it(file)
}
\arguments{
\item{file}{A file}
}
\description{
Load a data.frame or return NULL.
}
\author{
Marc Schwenzer
}
\keyword{internal}
