% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{reco_do}
\alias{reco_do}
\title{reco_do}
\usage{
reco_do(string, tab = TRUE)
}
\arguments{
\item{string}{a vector}

\item{a.df}{a data.frame}
}
\value{

}
\description{
reco_do
}
\details{
decide if ask_string_write_to_file is necessary
}
\author{
Marc Schwenzer
}
\keyword{internal}
