% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{reco_filters}
\alias{reco_filters}
\title{reco_filters}
\usage{
reco_filters(x)
}
\arguments{
\item{x}{a vector}
}
\value{
string modified by several possible filters
}
\description{
filter conditions
}
\author{
Marc Schwenzer
}
\keyword{internal}
