% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{set_a.df}
\alias{set_a.df}
\title{.. content for \description{} (no empty lines) ..}
\usage{
set_a.df(val)
}
\arguments{
\item{string}{}
}
\value{
Null
}
\description{
set_a.df
}
\details{
set a.df from reco env
}
\author{
Marc Schwenzer
}
