% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{st_to_df}
\alias{st_to_df}
\title{st_to_df}
\usage{
st_to_df(rules)
}
\arguments{
\item{rule}{}
}
\value{

}
\description{
st_to_df
}
\details{
Parse bracket-based rule specification
}
\author{
Marc Schwenzer
}
