% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{suggest_based_on_df}
\alias{suggest_based_on_df}
\title{suggest_based_on_df}
\usage{
suggest_based_on_df(string, tab = TRUE)
}
\description{
suggest_based_on_df
}
\details{
suggest_based_on_df
}
\author{
Marc Schwenzer
}
\keyword{internal}
