% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{validate_df_structure}
\alias{validate_df_structure}
\title{validate_df_structure}
\usage{
validate_df_structure(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
df
}
\description{
Check if the replacements has the correct format
}
\author{
Marc Schwenzer
}
\keyword{internal}
