% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{visualize_replacement_table}
\alias{visualize_replacement_table}
\title{visualize_replacement_table}
\usage{
visualize_replacement_table(
  replacements,
  type = "vertical",
  matched.values = matched.values
)
}
\arguments{
\item{replacements}{the replacements data.frame}

\item{type}{horizontal or vertical based on tab argument of reco}
}
\value{

}
\description{
visualize_replacement_table
}
\details{
Visualizer for replacemenent table based on stargazer
}
\author{
Marc Schwenzer
}
