% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{reco}
\alias{reco}
\title{reco}
\usage{
reco(string, file, interactive = TRUE, tab = TRUE)
}
\arguments{
\item{string}{The string to recode}

\item{file}{The file where the replacements are located. Columsn have to be named as from and to. If file is a data.frame, then this is used as replacement structure.}

\item{interactive}{If true an interactive coding session to add recodigs written to file, otherwise the recoding is done just base on the file leaving it as it is. The interactive coding session presents a table of alternatives. One can either enter a new string defining a new category to which the current value is assigned or enter a number that selects one of the previous categories. Entering 0 adds the current value as category. Entering '' (just return) ignores the current value.}

\item{tab}{Wheter to print the tab of alternatives/categories in every step.}
}
\value{
a vector.
}
\description{
Free and easy replace elements of a vector based on simple replacement files.
}
\author{
Marc Schwenzer
}
