% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{reco_replace_character}
\alias{reco_replace_character}
\title{reco_replace_character}
\usage{
reco_replace_character(string, a.df, a.df.class)
}
\arguments{
\item{string}{}

\item{a.df}{}
}
\value{

}
\description{
reco_replace_character
}
\details{
reco_replace_character
}
\author{
Marc Schwenzer
}
